add code comments to all classes
