+++ src/main/java/com/ddes/smartmeter/config/RabbitConfig.java
@@ -1,9 +1,13 @@
-package com.ddes.smartmeter.rabbit;
+package com.ddes.smartmeter.config;
 
 import org.springframework.amqp.core.Queue;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
+/**
+ * Configuration class for setting up RabbitMQ queues.
+ * This class defines beans that create and configure RabbitMQ queues for the application.
+ */
 @Configuration
 public class RabbitConfig {
