+++ src/main/resources/application.yml
@@ -1,17 +1,18 @@
 server:
+  port: 8443 # The port the application will run on. Default HTTPS port is 8443.
   ssl:
-    key-store: ${SMARTMETER_KEYSTORE_LOCATION}
-    key-store-password: ${SMARTMETER_KEYSTORE_PASSWORD}
-    key-store-type: JKS
-    trust-store: ${SMARTMETER_TRUSTSTORE_LOCATION}
-    trust-store-password: ${SMARTMETER_TRUSTSTORE_PASSWORD}
-    trust-store-type: JKS
+    key-store: ${SMARTMETER_KEYSTORE_LOCATION} # The path to the keystore file, see README for details.
+    key-store-password: ${SMARTMETER_KEYSTORE_PASSWORD} # THe password to access the keystore file, see README for details.
+    key-store-type: JKS # The type of keystore, set to Java KeyStore (JKS).
+    trust-store: ${SMARTMETER_TRUSTSTORE_LOCATION} # THe path to the truststore file, see README for details.
+    trust-store-password: ${SMARTMETER_TRUSTSTORE_PASSWORD} # THe password to access the truststore file, see README for details.
+    trust-store-type: JKS # The type of truststore, set to Java KeyStore (JKS).
 
 spring:
   application:
-    name: smartmeter
+    name: SmartMeter # The name of the Spring Boot application.
   rabbitmq:
-    host: ${SMARTMETER_RABBITMQ_HOST}
-    port: 5672
-    username: ${SMARTMETER_RABBITMQ_USERNAME}
-    password: ${SMARTMETER_RABBITMQ_PASSWORD}
+    host: ${SMARTMETER_RABBITMQ_HOST} # The hostname or IP address to connect to the RabbitMQ server.
+    port: 5672 # The port to connect to the RabbitMQ server. Default RabbitMQ port is 5672.
+    username: ${SMARTMETER_RABBITMQ_USERNAME} # The username for authenticating with the RabbitMQ server.
+    password: ${SMARTMETER_RABBITMQ_PASSWORD} # The password for the RabbitMQ user.
