+++ src/test/java/MeterReadingTests/MeterReadingTests.java
@@ -7,7 +7,6 @@ public class MeterReadingTests {
 
     @Test
     public void test01_WhenObjectBuilt_EnsureObjectNoExceptionsThrownAndValuesAreCorrect() {
-
         MeterReading meterReading = new MeterReading("123e4567-e89b-12d3-a456-426614174000", 10.0, 1234567890);
 
         assert(meterReading.getClientId().toString().equals("123e4567-e89b-12d3-a456-426614174000"));
@@ -15,6 +14,4 @@ public class MeterReadingTests {
         assert(meterReading.getCurrentCost() == 1.2);
         assert(meterReading.getTimestamp() == 1234567890);
     }
-
-
 }
