+++ src/test/java/OutageAlertServiceTests/OutageAlertServiceTests.java
@@ -15,5 +15,4 @@ public class OutageAlertServiceTests {
         ObjectNode jsonObject = outageAlertService.createAlert();
         assertEquals("Electricity grid is down.", jsonObject.get("message").asText());
     }
-
 }
