+++ src/test/java/RabbitServiceTests/RabbitServiceTests.java
@@ -1,7 +1,7 @@
-package RabbitReceiverTests;
+package RabbitServiceTests;
 
 import com.ddes.smartmeter.entities.MeterReading;
-import com.ddes.smartmeter.rabbit.RabbitReceiver;
+import com.ddes.smartmeter.services.RabbitService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.jupiter.api.Test;
 
@@ -9,9 +9,9 @@ import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class RabbitReceiverTests {
+public class RabbitServiceTests {
 
-    RabbitReceiver rabbitReceiver = new RabbitReceiver();
+    RabbitService rabbitService = new RabbitService();
 
     @Test
     public void test01_WhenCreateJsonObjectCalled_EnsureObjectBuiltWithCorrectValues() {
@@ -19,9 +19,10 @@ public class RabbitReceiverTests {
             UUID clientId = UUID.fromString("123e4567-e89b-12d3-a456-426614174000");
 
             MeterReading meterReading = new MeterReading(clientId.toString(), 10.0, 1234567890);
+
             double totalBill = 100.0;
 
-            String jsonString = rabbitReceiver.createJsonString(meterReading, totalBill);
+            String jsonString = rabbitService.createJsonString(meterReading, totalBill);
 
             assertEquals("{\"clientId\":\"123e4567-e89b-12d3-a456-426614174000\",\"currentUsage\":10.0,\"currentCost\":1.2,\"totalBill\":100.0,\"timestamp\":1234567890}", jsonString);
         } catch (JsonProcessingException e) {
@@ -35,14 +36,14 @@ public class RabbitReceiverTests {
             UUID clientId = UUID.fromString("123e4567-e89b-12d3-a456-426614174000");
 
             MeterReading meterReading = new MeterReading(clientId.toString(), -10.0, 1234567890);
+
             double totalBill = 100.0;
 
-            String jsonString = rabbitReceiver.createJsonString(meterReading, totalBill);
+            String jsonString = rabbitService.createJsonString(meterReading, totalBill);
 
             assertEquals("{\"clientId\":\"123e4567-e89b-12d3-a456-426614174000\",\"currentUsage\":-10.0,\"currentCost\":-1.2,\"totalBill\":100.0,\"timestamp\":1234567890}", jsonString);
         } catch (JsonProcessingException e) {
             e.printStackTrace();
         }
     }
-
 }
